+++ projects/ngx-scrollbar-demo/src/app/home-page/home-page.component.ts
@@ -1,12 +1,14 @@
 import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { Platform } from '@angular/cdk/platform';
+import { MatButton } from '@angular/material/button';
+import { NgScrollbarExtDirective, ScrollbarDocument } from 'ngx-scrollbar';
 import { Example2Component } from '../example2/example2.component';
 import { Example3Component } from '../example3/example3.component';
 import { Example5Component } from '../example5/example5.component';
 import { LabComponent } from '../lab/lab.component';
 import { HeaderComponent } from '../shared/header/header.component';
-import { MatButton } from '@angular/material/button';
-import { NgScrollbar, NgScrollbarExt, ScrollbarAnywhere, ScrollbarRef } from 'ngx-scrollbar';
+import { AsyncDetection, NgScrollbar, NgScrollbarExt, ScrollbarAnywhere, ScrollbarRef } from 'ngx-scrollbar';
+import { NgScrollReachDrop } from 'ngx-scrollbar/reached-event';
 
 @Component({
   selector: 'app-home-page',
@@ -20,16 +22,35 @@ import { NgScrollbar, NgScrollbarExt, ScrollbarAnywhere, ScrollbarRef } from 'ng
     HeaderComponent,
     MatButton,
     NgScrollbarExt,
+    AsyncDetection,
+    NgScrollbarExtDirective,
+    NgScrollReachDrop,
   ],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class HomePageComponent {
+  externalViewport = '.my-custom-viewport';
+  externalContentWrapper = '.my-custom-content-wrapper';
+  externalSpacer = '.my-custom-spacer';
+
+  show = false;
+
   readonly platform: Platform = inject(Platform);
 
   readonly anyWhere = inject(ScrollbarAnywhere);
+  readonly scrollbarDocument = inject(ScrollbarDocument);
 
   flag = false;
+  reachedDisabled = false;
   scrollbarRef: ScrollbarRef<NgScrollbar>;
+
+  x(t: NgScrollbarExt) {
+    console.log('🤖', t);
+  }
+
+  log(e){
+    console.log('🐴', e);
+  }
 
   test() {
     if (this.flag) {
@@ -42,7 +63,10 @@ export class HomePageComponent {
 
     requestAnimationFrame(() => {
       this.scrollbarRef = this.anyWhere.createScrollbar('#test');
-    })
+    });
+    // requestAnimationFrame(() => {
+    //   this.scrollbarDocument.attachScrollbar();
+    // });
 
     // requestAnimationFrame(() => {
     //   this.anyWhere.createScrollbar({