+++ projects/ngx-scrollbar/src/lib/async-detection.ts
@@ -2,9 +2,7 @@ import {
   Directive,
   inject,
   effect,
-  untracked,
   input,
-  NgZone,
   InputSignal,
   EffectCleanupRegisterFn
 } from '@angular/core';
@@ -12,69 +10,70 @@ import { ContentObserver } from '@angular/cdk/observers';
 import { Subscription, throttleTime } from 'rxjs';
 import { NgScrollbarExt } from './ng-scrollbar-ext';
 
+type AsyncDetectionOption = '' | 'auto';
+
+// TODO: Rename the class to ScrollbarAsyncViewport
 @Directive({
   selector: 'ng-scrollbar[externalViewport][asyncDetection]'
 })
 export class AsyncDetection {
 
   private readonly scrollbar: NgScrollbarExt = inject(NgScrollbarExt, { self: true });
-  private readonly zone: NgZone = inject(NgZone);
 
   private readonly contentObserver: ContentObserver = inject(ContentObserver);
 
-  asyncDetection: InputSignal<'auto' | ''> = input.required<'auto' | ''>();
+  asyncDetection: InputSignal<AsyncDetectionOption> = input.required<AsyncDetectionOption>();
 
   constructor() {
     this.scrollbar.skipInit = true;
     let sub$: Subscription;
+    let init: boolean;
 
     effect((onCleanup: EffectCleanupRegisterFn) => {
-      const init: boolean = this.scrollbar.viewport.initialized();
       const externalViewport: string = this.scrollbar.externalViewport();
       const externalContentWrapper: string = this.scrollbar.externalContentWrapper();
       const externalSpacer: string = this.scrollbar.externalSpacer();
-      const asyncDetection: 'auto' | '' = this.asyncDetection();
+      const asyncDetection: AsyncDetectionOption = this.asyncDetection();
 
-      untracked(() => {
-        let viewportElement: HTMLElement;
-        let contentWrapperElement: HTMLElement;
+      // The content observer should not be throttled using the same function we use for ResizeObserver,
+      // It should detect the content change asap to attach the scrollbar
+      sub$ = this.contentObserver.observe(this.scrollbar.nativeElement).pipe(
+        throttleTime(100, null, {
+          leading: true,
+          trailing: true
+        })
+      ).subscribe(() => {
+        // Search for the viewport element
+        this.scrollbar.viewportElement.set(this.scrollbar.getElement(externalViewport))
+        // Search for the content wrapper element
+        this.scrollbar.contentWrapperElement.set(this.scrollbar.getElement(externalContentWrapper));
+        // Search for the spacer element
+        this.scrollbar.spacerElement.set(this.scrollbar.getElement(externalSpacer));
 
-        this.zone.runOutsideAngular(() => {
-          // The content observer should not be throttled using the same function we use for ResizeObserver,
-          // It should detect the content change asap to attach the scrollbar
-          sub$ = this.contentObserver.observe(this.scrollbar.nativeElement).pipe(
-            throttleTime(100, null, {
-              leading: true,
-              trailing: true
-            })
-          ).subscribe(() => {
-            // Search for external viewport
-            viewportElement = this.scrollbar.nativeElement.querySelector(externalViewport);
+        const contentWrapperCheck: boolean = externalContentWrapper ? !!this.scrollbar.contentWrapperElement() : true;
+        const spacerPassCheck: boolean = externalSpacer ? !!this.scrollbar.spacerElement() : true;
 
-            // Search for external content wrapper
-            contentWrapperElement = this.scrollbar.nativeElement.querySelector(externalContentWrapper);
+        if (!init && this.scrollbar.viewportElement() && contentWrapperCheck && spacerPassCheck) {
+          // If an external spacer selector is provided, search for it
+          this.scrollbar.initialize(
+            this.scrollbar.viewportElement(),
+            this.scrollbar.contentWrapperElement(),
+            this.scrollbar.spacerElement()
+          );
+          init = true;
+        } else if (!this.scrollbar.viewportElement() ||
+          (externalContentWrapper && !this.scrollbar.contentWrapperElement()) ||
+          (externalSpacer && !this.scrollbar.spacerElement())) {
+          this.scrollbar.destroy();
+          init = false;
+        }
 
-            this.zone.run(() => {
-              if (!init && viewportElement && contentWrapperElement) {
-                // If an external spacer selector is provided, search for it
-                let spacerElement: HTMLElement;
-                if (externalSpacer) {
-                  spacerElement = this.scrollbar.nativeElement.querySelector(externalSpacer);
-                }
-                this.scrollbar.initialize(viewportElement, contentWrapperElement, spacerElement);
-              } else if (!viewportElement || !contentWrapperElement) {
-                this.scrollbar.viewport.reset();
-              }
-            });
-
-            if (asyncDetection !== 'auto') {
-              sub$.unsubscribe();
-            }
-          });
-        });
-
-        onCleanup(() => sub$?.unsubscribe());
+        if (asyncDetection !== 'auto') {
+          sub$.unsubscribe();
+        }
       });
+
+      onCleanup(() => sub$?.unsubscribe());
     });
   }
 }